requirements.txt
asgiref==3.2.10
asyncio==3.4.3
blockcypher==1.0.90
boto3==1.17.3
botocore==1.20.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-session-csrf==0.7.1
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
marshmallow==3.11.1
mysql-connector-python==8.0.21
mysqlclient==2.1.0
pkg-resources==0.0.0
protobuf==3.13.0
pycparser==2.20
PyJWT==1.7.1
pyotp==2.4.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10